#AWARD-176 add additonal input types

+++ src/SFA.DAS.AODP.Application/Commands/Application/Page/UpdatePageAnswersCommand.cs
@@ -31,7 +31,7 @@ public class UpdatePageAnswersCommand : IRequest<BaseMediatrResponse<EmptyRespon
     public class Answer
     {
         public string? TextValue { get; set; }
-        public double? NumberValue { get; set; }
+        public decimal? NumberValue { get; set; }
         public DateTime? DateValue { get; set; }
         public List<string>? MultipleChoiceValue { get; set; }
         public string? RadioChoiceValue { get; set; }
